requirements.txt
django
djangorestframework
environs
psycopg2
pandas
celery
rabbitmq
opencv-python
flower
